Extract the automatic ordering toggle into its own file
